Add first test for #3. Refactor Uuid trait for model use.
