+++ database/factories/PartyFactory.php
@@ -1,0 +1,9 @@
+<?php
+
+use Faker\Generator as Faker;
+
+$factory->define(\App\Party::class, function (Faker $faker) {
+    return [
+        //
+    ];
+});
