+++ resources/assets/js/app.js
@@ -14,8 +14,6 @@ if (token) {
   console.error('CSRF token not found: https://laravel.com/docs/csrf#csrf-x-csrf-token');
 }
 
-// require('./echo');
-
 /**
  * Setup Vue for the project
  */
